fixup: Reduce test code duplication
Remove some test code utilities which had be added to multiple mldsa
test files.  This reduces code duplication in the test infrastructure
and simplifies maintainence.
