+++ runtime/tests/runtime_integration_tests/test_sign_with_exported_mldsa.rs
@@ -3,8 +3,8 @@
 //! Integration tests for the SIGN_WITH_EXPORTED_MLDSA mailbox command.
 //!
 //! An exported ML-DSA CDI slot is produced by INVOKE_DPE_MLDSA87 carrying a DPE
-//! `DeriveContext{EXPORT_CDI | CREATE_CERTIFICATE}` (see `export_cdi` below), so
-//! the successful-sign lineage is exercised here alongside the
+//! `DeriveContext{EXPORT_CDI | CREATE_CERTIFICATE}` (see `common::export_mldsa_cdi`),
+//! so the successful-sign lineage is exercised here alongside the
 //! request-validation paths (privilege, malformed request, handle-not-found)
 //! that run before the CDI lookup.
 
@@ -12,29 +12,18 @@ use caliptra_api::SocManager;
 use caliptra_builder::firmware::APP_MLDSA_ATTESTATION;
 use caliptra_common::checksum::calc_checksum;
 use caliptra_common::mailbox_api::{
-    CommandId, MailboxReq, MailboxReqHeader, PopulatePqCertReq, RevokeExportedCdiHandleReq,
-    SetPqSeedReq, SignWithExportedMldsaReq, SignWithExportedMldsaResp, SET_PQ_SEED_SEED_SIZE,
+    CommandId, MailboxReq, MailboxReqHeader, RevokeExportedCdiHandleReq, SignWithExportedMldsaReq,
+    SignWithExportedMldsaResp,
 };
-use caliptra_drivers::{Mldsa87Signature, MLDSA87_MU_BYTES};
+use caliptra_drivers::MLDSA87_MU_BYTES;
 use caliptra_error::CaliptraError;
 use caliptra_hw_model::{DefaultHwModel, HwModel, ModelError};
-use caliptra_runtime::{CaliptraDpeProfile, RtBootStatus};
-use caliptra_x509::MlDsa87CertBuilder;
-use dpe::{
-    commands::{Command, DeriveContextCmd, DeriveContextFlags},
-    context::ContextHandle,
-    response::Response,
-    tci::TciMeasurement,
-    TCI_SIZE,
-};
-use openssl::pkey::Private;
-use openssl::pkey_ctx::PkeyCtx;
-use openssl::pkey_ml_dsa::{PKeyMlDsaBuilder, Variant};
-use openssl::signature::Signature;
+use caliptra_runtime::RtBootStatus;
 use zerocopy::{FromBytes, IntoBytes};
 
 use crate::common::{
-    assert_error, execute_dpe_cmd, run_pqc_rt_test, run_rt_test, DpeResult, RuntimeTestArgs,
+    assert_error, export_mldsa_cdi, populate_pq_cert, provision_pq_seed, run_pqc_rt_test,
+    run_rt_test, RuntimeTestArgs,
 };
 
 /// Issue SIGN_WITH_EXPORTED_MLDSA and return the raw response bytes.
@@ -89,74 +78,6 @@ fn sign_full(
         .map(|resp| resp.expect("expected a SIGN_WITH_EXPORTED_MLDSA response"))
 }
 
-/// Provision the PQ.DevID CDI via SET_PQ_SEED, enabling PQC mode. Signing with an
-/// exported CDI requires this to have run first.
-fn set_pq_seed(model: &mut DefaultHwModel) {
-    let mut cmd = MailboxReq::SetPqSeed(SetPqSeedReq {
-        hdr: MailboxReqHeader { chksum: 0 },
-        seed: [0x5a; SET_PQ_SEED_SEED_SIZE],
-    });
-    cmd.populate_chksum().unwrap();
-    model
-        .mailbox_execute(u32::from(CommandId::SET_PQ_SEED), cmd.as_bytes().unwrap())
-        .expect("SET_PQ_SEED failed");
-}
-
-/// Populate the ML-DSA PQ certificate. Exporting a CDI with a certificate
-/// requires the PQ cert to have been populated first.
-fn populate_pq_cert(model: &mut DefaultHwModel) {
-    // Generate an ML-DSA-87 key pair and self-sign a placeholder TBS.
-    let pk_builder = PKeyMlDsaBuilder::<Private>::from_seed(Variant::MlDsa87, &[0u8; 32]).unwrap();
-    let priv_key = pk_builder.build().unwrap();
-    let tbs: &[u8] = b"this is going to be the TBS";
-    let mut sig_bytes = vec![];
-    let mut ctx = PkeyCtx::new(&priv_key).unwrap();
-    let mut algo = Signature::for_ml_dsa(Variant::MlDsa87).unwrap();
-    ctx.sign_message_init(&mut algo).unwrap();
-    ctx.sign_to_vec(tbs, &mut sig_bytes).unwrap();
-    let sig = Mldsa87Signature::new(sig_bytes.try_into().unwrap());
-    let builder = MlDsa87CertBuilder::new(tbs, &sig).unwrap();
-    let mut cert = [0u8; PopulatePqCertReq::MAX_CERT_SIZE];
-    let cert_size = builder.build(&mut cert).unwrap();
-
-    let mut cmd = MailboxReq::PopulatePqCert(PopulatePqCertReq {
-        hdr: MailboxReqHeader { chksum: 0 },
-        cert_size: cert_size as u32,
-        cert,
-    });
-    cmd.populate_chksum().unwrap();
-    model
-        .mailbox_execute(
-            u32::from(CommandId::POPULATE_PQ_CERT),
-            cmd.as_bytes().unwrap(),
-        )
-        .expect("POPULATE_PQ_CERT failed");
-}
-
-/// Export an ML-DSA CDI via INVOKE_DPE_MLDSA87 + `DeriveContext{EXPORT_CDI}` and
-/// return the exported-CDI handle that SIGN/REVOKE consume. Requires SET_PQ_SEED
-/// and POPULATE_PQ_CERT to have run first.
-fn export_cdi(model: &mut DefaultHwModel) -> [u8; 32] {
-    let derive_ctx_cmd = DeriveContextCmd {
-        handle: ContextHandle::default(),
-        data: TciMeasurement([0; TCI_SIZE]),
-        flags: DeriveContextFlags::EXPORT_CDI | DeriveContextFlags::CREATE_CERTIFICATE,
-        tci_type: 0,
-        target_locality: 0,
-        ..Default::default()
-    };
-    let resp = execute_dpe_cmd(
-        CaliptraDpeProfile::Mldsa,
-        model,
-        &mut Command::DeriveContext(&derive_ctx_cmd),
-        DpeResult::Success,
-    );
-    let Some(Response::DeriveContextExportedCdi(resp)) = resp else {
-        panic!("expected derive context exported cdi resp!");
-    };
-    resp.header.exported_cdi
-}
-
 /// Issue REVOKE_EXPORTED_CDI_HANDLE for `handle`, expecting success.
 fn revoke_cdi(model: &mut DefaultHwModel, handle: [u8; 32]) {
     let mut cmd = MailboxReq::RevokeExportedCdiHandle(RevokeExportedCdiHandleReq {
@@ -175,9 +96,9 @@ fn revoke_cdi(model: &mut DefaultHwModel, handle: [u8; 32]) {
 #[test]
 fn test_sign_with_exported_mldsa_success_sign_data() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
     populate_pq_cert(&mut model);
-    let handle = export_cdi(&mut model);
+    let handle = export_mldsa_cdi(&mut model);
 
     // Data mode: the firmware signs the raw message directly.
     let message = b"caliptra exported ml-dsa sign-data test message";
@@ -200,9 +121,9 @@ fn test_sign_with_exported_mldsa_success_sign_data() {
 #[test]
 fn test_sign_with_exported_mldsa_success_sign_external_mu() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
     populate_pq_cert(&mut model);
-    let handle = export_cdi(&mut model);
+    let handle = export_mldsa_cdi(&mut model);
 
     // External-mu mode: the caller supplies the 64-byte mu and the firmware signs
     // it directly.
@@ -226,11 +147,11 @@ fn test_sign_with_exported_mldsa_success_sign_external_mu() {
 #[test]
 fn test_sign_with_exported_mldsa_wrong_handle_after_export() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
     populate_pq_cert(&mut model);
     // A CDI is exported, but a handle that does not match the active slot must
     // still be rejected as not found.
-    let _handle = export_cdi(&mut model);
+    let _handle = export_mldsa_cdi(&mut model);
 
     let result = sign(
         &mut model,
@@ -248,9 +169,9 @@ fn test_sign_with_exported_mldsa_wrong_handle_after_export() {
 #[test]
 fn test_sign_with_exported_mldsa_sign_after_revoke() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
     populate_pq_cert(&mut model);
-    let handle = export_cdi(&mut model);
+    let handle = export_mldsa_cdi(&mut model);
 
     // Signing succeeds while the exported-CDI slot is active.
     sign(
@@ -281,9 +202,9 @@ fn test_sign_with_exported_mldsa_sign_after_revoke() {
 #[test]
 fn test_sign_with_exported_mldsa_sign_after_disable_attestation() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
     populate_pq_cert(&mut model);
-    let handle = export_cdi(&mut model);
+    let handle = export_mldsa_cdi(&mut model);
 
     // Signing succeeds before attestation is disabled.
     sign(
@@ -323,7 +244,7 @@ fn test_sign_with_exported_mldsa_sign_after_disable_attestation() {
 #[test]
 fn test_sign_with_exported_mldsa_invalid_sign_mode() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
 
     // sign_mode is neither SIGN_MODE_DATA nor SIGN_MODE_EXTERNAL_MU.
     let result = sign(&mut model, [0u8; 32], 0xDEAD_BEEF, b"message");
@@ -337,7 +258,7 @@ fn test_sign_with_exported_mldsa_invalid_sign_mode() {
 #[test]
 fn test_sign_with_exported_mldsa_external_mu_wrong_size() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
 
     // External-mu mode requires exactly MLDSA87_MU_BYTES (64) of message; any
     // other length is rejected before the CDI lookup.
@@ -360,7 +281,7 @@ fn test_sign_with_exported_mldsa_external_mu_wrong_size() {
 #[test]
 fn test_sign_with_exported_mldsa_message_too_large() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
 
     // A message_size larger than the buffer must be rejected as invalid params
     // (and must not be used to index the message buffer).
@@ -379,7 +300,7 @@ fn test_sign_with_exported_mldsa_message_too_large() {
 #[test]
 fn test_sign_with_exported_mldsa_handle_not_found() {
     let mut model = run_pqc_rt_test();
-    set_pq_seed(&mut model);
+    provision_pq_seed(&mut model);
 
     // No CDI has been exported, so any handle must be rejected as not found.
     let result = sign(
